refactor: splitted chart code  for a util and component
