+++ frontend/src/components/molecules/PowerUnitEngineChart.tsx
@@ -1,150 +1,7 @@
-import React, { useEffect } from "react";
+import { useEffect } from "react";
 import Plotly from "plotly.js-dist-min";
 import { useEngineStore } from "../../utils/useEngine";
-import { density } from "../../utils/atmosphere";
-
-interface powerInput {
-  seaLevelPower: number;
-  engineSpeed: number;
-  reductionRatio: number;
-  maxAltitude: number;
-  kCoefficient: number;
-  halfSupercharger?: { endAltitude: number; endPower: number };
-  endSupercharger?: {
-    startAltitude: number;
-    endAltitude: number;
-    endPower: number;
-  };
-}
-const altitudeArrayCreator = (
-  maxAltitude: number,
-  halfSuperchargerendAltitude?: number,
-  endSuperchargerendAltitude?: number
-) => {
-  let altitudes = Array.from(
-    { length: Math.floor(maxAltitude / 100) },
-    (_, i) => 100 * i
-  );
-  altitudes = [...altitudes, maxAltitude];
-
-  if (
-    halfSuperchargerendAltitude &&
-    altitudes.includes(halfSuperchargerendAltitude) === false
-  ) {
-    altitudes = [...altitudes, halfSuperchargerendAltitude];
-    altitudes.sort(function (a, b) {
-      return a - b;
-    });
-  }
-
-  if (
-    endSuperchargerendAltitude &&
-    altitudes.includes(endSuperchargerendAltitude) === false
-  ) {
-    altitudes = [...altitudes, endSuperchargerendAltitude];
-    altitudes.sort(function (a, b) {
-      return a - b;
-    });
-  }
-  return altitudes;
-};
-const superchargerPower = (
-  altitude: number,
-  superchargerStartAltitude: number,
-  superchargerEndAltitude: number,
-  superchargerStartPower: number,
-  superchargerEndPower: number
-) => {
-  return (
-    superchargerStartPower +
-    ((superchargerEndPower - superchargerStartPower) /
-      (superchargerEndAltitude - superchargerStartAltitude)) *
-      (altitude - superchargerStartAltitude)
-  );
-};
-const normalPower = (
-  altitude: number,
-  startAltitude: number,
-  startPower: number,
-  kCoefficient: number
-) => {
-  return (
-    (startPower * (density(altitude) / density(startAltitude) - kCoefficient)) /
-    (1 - kCoefficient)
-  );
-};
-const powerArrayCreator = (altitude: number, props: powerInput) => {
-  let power = normalPower(altitude, 0, props.seaLevelPower, props.kCoefficient);
-
-  if (
-    props.halfSupercharger &&
-    altitude <= props.halfSupercharger.endAltitude
-  ) {
-    power = superchargerPower(
-      altitude,
-      0,
-      props.halfSupercharger.endAltitude,
-      props.seaLevelPower,
-      props.halfSupercharger.endPower
-    );
-  } else if (
-    props.halfSupercharger &&
-    altitude > props.halfSupercharger.endAltitude
-  ) {
-    power = normalPower(
-      altitude,
-      props.halfSupercharger.endAltitude,
-      props.halfSupercharger.endPower,
-      props.kCoefficient
-    );
-  }
-
-  if (
-    props.halfSupercharger &&
-    props.endSupercharger &&
-    altitude >= props.endSupercharger.startAltitude &&
-    altitude <= props.endSupercharger.endAltitude
-  ) {
-    let startPower = normalPower(
-      props.endSupercharger.startAltitude,
-      props.halfSupercharger.endAltitude,
-      props.halfSupercharger.endPower,
-      props.kCoefficient
-    );
-    power = superchargerPower(
-      altitude,
-      props.endSupercharger.startAltitude,
-      props.endSupercharger.endAltitude,
-      startPower,
-      props.endSupercharger.endPower
-    );
-  } else if (
-    props.endSupercharger &&
-    altitude > props.endSupercharger.endAltitude
-  ) {
-    power = normalPower(
-      altitude,
-      props.endSupercharger.endAltitude,
-      props.endSupercharger.endPower,
-      props.kCoefficient
-    );
-  }
-
-  return power;
-};
-const powerFunction = (props: powerInput) => {
-  let altitudes = altitudeArrayCreator(
-    props.maxAltitude,
-    props.halfSupercharger?.endAltitude,
-    props.endSupercharger?.endAltitude
-  );
-
-  console.log(altitudes);
-  let power = Array.from(altitudes, (altitude) =>
-    powerArrayCreator(altitude, props)
-  );
-  return { x: altitudes, y: power };
-};
+import { powerFunction } from "../../utils/enginePower";
 
 let layout = {
   title: {
@@ -172,25 +29,18 @@ let layout = {
 };
 
 const PowerUnitEngineChart = () => {
-  const seaLevelPower = useEngineStore((state) => state.seaLevelPower);
-  const engineSpeed = useEngineStore((state) => state.engineSpeed);
-  const reductionRatio = useEngineStore((state) => state.reductionRatio);
-  const maxAltitude = useEngineStore((state) => state.maxAltitude);
-  const kCoefficient = useEngineStore((state) => state.kCoefficient);
-
+  // const seaLevelPower = useEngineStore((state) => state.seaLevelPower);
+  // const engineSpeed = useEngineStore((state) => state.engineSpeed);
+  // const reductionRatio = useEngineStore((state) => state.reductionRatio);
+  // const maxAltitude = useEngineStore((state) => state.maxAltitude);
+  // const kCoefficient = useEngineStore((state) => state.kCoefficient);
+  const props = useEngineStore();
+  
   useEffect(() => {
-    const trace = powerFunction({
-      seaLevelPower: 800,
-      engineSpeed: 3000,
-      reductionRatio: 0.4,
-      maxAltitude: 10010,
-      kCoefficient: 0.1,
-      halfSupercharger: { endAltitude: 3010, endPower: 1000 },
-      endSupercharger: {startAltitude:5000 , endAltitude:8000 , endPower:900}
-    });
+    let trace = powerFunction(props);
     console.log(trace);
     Plotly.newPlot("plot", [trace], layout);
-  }, []);
+  });
 
   return (
     <div className=" w-full text-bold text-black text-3xl flex flex-col justify-center items-center">
