+++ frontend/src/utils/useEngine.ts
@@ -1,18 +1,18 @@
-import create from 'zustand'
+import create from "zustand";
 
 interface EngineState {
-  seaLevelPower: number
-  engineSpeed: number
-  reductionRatio: number
-  maxAltitude: number
-  kCoefficient: number
-  engineType: string
-  setSeaLevelPower: (value: number) => void
-  setEngineSpeed: (value: number) => void
-  setReductionRatio: (value: number) => void
-  setMaxAltitude: (value: number) => void
-  setKCoefficient: (value: number) => void
-  setEngineType: (value: number) => void
+  seaLevelPower: number;
+  engineSpeed: number;
+  reductionRatio: number;
+  maxAltitude: number;
+  kCoefficient: number;
+  engineType: string;
+  setSeaLevelPower: (value: number) => void;
+  setEngineSpeed: (value: number) => void;
+  setReductionRatio: (value: number) => void;
+  setMaxAltitude: (value: number) => void;
+  setKCoefficient: (value: number) => void;
+  setEngineType: (value: number) => void;
 }
 
 export const useEngineStore = create<EngineState>()((set) => ({
@@ -25,7 +25,7 @@ export const useEngineStore = create<EngineState>()((set) => ({
   setSeaLevelPower: (value) => set((state) => ({ seaLevelPower: value })),
   setEngineSpeed: (value) => set((state) => ({ engineSpeed: value })),
   setReductionRatio: (value) => set((state) => ({ reductionRatio: value })),
-  setMaxAltitude:(value) => set((state) => ({maxAltitude: value})),
-  setKCoefficient:(value) => set((state) => ({maxAltitude: value})),
-  setEngineType:(value) => set((state) => ({maxAltitude: value}))
-}))
+  setMaxAltitude: (value) => set((state) => ({ maxAltitude: value })),
+  setKCoefficient: (value) => set((state) => ({ maxAltitude: value })),
+  setEngineType: (value) => set((state) => ({ maxAltitude: value })),
+}));
